+++ KubePrompt/Init/Init.ps1
@@ -1,0 +1,15 @@
+$global:KubePromptSettings = @{
+    SymbolColor = ([ConsoleColor]::DarkBlue)
+    ContextColor = ([ConsoleColor]::DarkRed)
+    NamespaceColor = ([ConsoleColor]::DarkCyan)
+    Prefix = '['
+    Suffix = ']'
+    Symbol = @{
+        Enabled = $true
+        Value = 0x2388
+    }
+    Divider = ':'
+    Separator = '|'
+}
+
+Install-PromptOverrideIfDefault